Update multiples.cpp
Here is the setup.
You’re a backend developer and you need to write a code that outputs an integer, a multiple of 10, between 0 & 100 given a rational number as input, a number between 0 & 1.
Purpose: That number is used to reference a css rule used to generate a graphic that represents a sliding rating bar (see image below).  An example of input would be 0.8333 and your code needs to output 80.

Rules for code
Your code needs to output one of the following whole numbers: 10, 20, .. ,.90, and 100 given the following.
1.	I need to be able to enter in a rational whole number between 0 & 1.

2.	Your should produce two bits of code implementing:
a.	An IF or Switch statement (whichever you prefer).
b.	Your own elegant solution that uses a single line to produce the output. 
i.	Meaning, an If and switch statement uses multiple lines of conditional test to produce a result.  This same operation can be done on a single line and you need to find that solution. 

NOTE: you cannot create a function and call it from a single line.  There is a built in mechanism that allows you to produce the same result that the IF or Switch statement generates. 

3.	For all methods the following inputs should generate the respective outputs:
a.	0.14 outputs 10
b.	0.15 outputs 20
c.	0.41 outputs 40
d.	0.49 outputs 50

Notice that the hundredth decimal place determines which direction on the number line to go.  A hundredth place greater than or equal to 5 and less than or equal to 9 produce an output that moves right on the number line, while a hundredth place less than or equal to 4 or greater than or equal to 0 produces an output that moves left on the number line.